refactor : 의존관계 정리
